requirements.txt
biopython==1.84
python-dotenv==1.0.1
tqdm==4.66.5

chromadb==0.5.3
sentence-transformers==3.0.1

openai==2.24.0
httpx==0.27.2